Handle json and regular printing for all commands in their own respective files for readability
