+++ .gitignore
@@ -1,2 +1,3 @@
 cover.out
 mc
+*~
